+++ jl/math_libm.jl
@@ -91,8 +91,8 @@ fpart(x) = x - trunc(x)
 #abs(x::Float32) = ccall(dlsym(_jl_libfdm, :fabsf), Float32, (Float32,), x)
 #@vectorize_1arg Number abs
 
-lgamma(x::Float64) = ccall(dlsym(_jl_libamos,:dgamln_),Float64,(Ptr{Float64},Ptr{Int32}),&x,&int32(0))
-lgamma(x::Float32) = ccall(dlsym(_jl_libamos,:gamln_),Float32,(Ptr{Float32},Ptr{Int32}),&x,&int32(0))
+lgamma(x::Float64) = ccall(dlsym(_jl_libamos,:dgamln_),Float64,(Ptr{Float64},Ptr{Int32}),&x,&0)
+lgamma(x::Float32) = ccall(dlsym(_jl_libamos,:gamln_),Float32,(Ptr{Float32},Ptr{Int32}),&x,&0)
 lgamma(x::Real) = lgamma(float(x))
 @vectorize_1arg Number lgamma
 
@@ -184,7 +184,7 @@ function airy(k::Int, z::Complex128)
               (Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32},
                Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32}),
               &real(z), &imag(z),
-              &id, &int32(1),
+              &id, &1,
               pointer(ai,1), pointer(ai,2),
               pointer(ae,1), pointer(ae,2))
         return complex(ai[1],ai[2])
@@ -193,7 +193,7 @@ function airy(k::Int, z::Complex128)
               (Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32},
                Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32}),
               &real(z), &imag(z),
-              &id, &int32(1),
+              &id, &1,
               pointer(ai,1), pointer(ai,2),
               pointer(ae,1), pointer(ae,2))
         return complex(ai[1],ai[2])
@@ -218,7 +218,7 @@ let
         ccall(dlsym(_jl_libamos, :zbesh_), Void,
               (Ptr{Float64}, Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32},
                Ptr{Int32}, Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32}),
-              &real(z), &imag(z), &nu, &int32(1), &int32(k), &int32(1),
+              &real(z), &imag(z), &nu, &1, &k, &1,
               pointer(cy,1), pointer(cy,2),
               pointer(ae,1), pointer(ae,2))
         return complex(cy[1],cy[2])
@@ -228,7 +228,7 @@ let
         ccall(dlsym(_jl_libamos, :zbesi_), Void,
               (Ptr{Float64}, Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32},
                Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32}),
-              &real(z), &imag(z), &nu, &int32(1), &int32(1),
+              &real(z), &imag(z), &nu, &1, &1,
               pointer(cy,1), pointer(cy,2),
               pointer(ae,1), pointer(ae,2))
         return complex(cy[1],cy[2])
@@ -238,7 +238,7 @@ let
         ccall(dlsym(_jl_libamos, :zbesj_), Void,
               (Ptr{Float64}, Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32},
                Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32}),
-              &real(z), &imag(z), &nu, &int32(1), &int32(1),
+              &real(z), &imag(z), &nu, &1, &1,
               pointer(cy,1), pointer(cy,2),
               pointer(ae,1), pointer(ae,2))
         return complex(cy[1],cy[2])
@@ -248,7 +248,7 @@ let
         ccall(dlsym(_jl_libamos, :zbesk_), Void,
               (Ptr{Float64}, Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32},
                Ptr{Float64}, Ptr{Float64}, Ptr{Int32}, Ptr{Int32}),
-              &real(z), &imag(z), &nu, &int32(1), &int32(1),
+              &real(z), &imag(z), &nu, &1, &1,
               pointer(cy,1), pointer(cy,2),
               pointer(ae,1), pointer(ae,2))
         return complex(cy[1],cy[2])
@@ -259,7 +259,7 @@ let
               (Ptr{Float64}, Ptr{Float64}, Ptr{Float64}, Ptr{Int32},
                Ptr{Int32}, Ptr{Float64}, Ptr{Float64}, Ptr{Int32},
                Ptr{Float64}, Ptr{Float64}, Ptr{Int32}),
-              &real(z), &imag(z), &nu, &int32(1), &int32(1),
+              &real(z), &imag(z), &nu, &1, &1,
               pointer(cy,1), pointer(cy,2),
               pointer(ae,1), pointer(wrk,1),
               pointer(wrk,2), pointer(ae,2))
